default avatarURL is now set for newly created user, and added logic to prevent deletion of default image #176

+++ backend/src/main/java/com/capstone/moneytree/service/impl/DefaultUserService.java
@@ -37,14 +37,16 @@ public class DefaultUserService implements UserService {
    private final AmazonS3Service amazonS3Service;
 
    private final String bucketName;
+   private final String defaultProfileURL;
 
    @Autowired
-   public DefaultUserService(UserDao userDao, ValidatorFactory validatorFactory, AmazonS3Service amazonS3Service, @Value("${aws.profile.pictures.bucket}") String bucketName) {
+   public DefaultUserService(UserDao userDao, ValidatorFactory validatorFactory, AmazonS3Service amazonS3Service, @Value("${aws.profile.pictures.bucket}") String bucketName, @Value("${aws.default.profile.picture}") String defaultProfileURL) {
       this.userDao = userDao;
       this.validatorFactory = validatorFactory;
       this.passwordEncryption = new MoneyTreePasswordEncryption();
       this.amazonS3Service = amazonS3Service;
       this.bucketName = bucketName;
+      this.defaultProfileURL = defaultProfileURL;
    }
 
    @Override
@@ -78,6 +80,7 @@ public class DefaultUserService implements UserService {
       String password = user.getPassword();
       String encryptedPassword = encryptData(password);
       user.setPassword(encryptedPassword);
+      user.setAvatarURL(String.format("https://%s.s3.amazonaws.com/%s", bucketName, defaultProfileURL));
 
       userDao.save(user);
 
@@ -91,8 +94,8 @@ public class DefaultUserService implements UserService {
       //since user exists, we can now upload image to s3 and save imageUrl into db
       String imageUrl = amazonS3Service.uploadImageToS3Bucket(imageFile, getBucketName());
 
-      //if user already has a profile picture, handle deleting old picture
-      if (StringUtils.isNotBlank(user.getAvatarURL())) {
+      //if user already has a profile picture that is not the default picture, handle deleting old picture
+      if (StringUtils.isNotBlank(user.getAvatarURL()) && !user.getAvatarURL().contains(defaultProfileURL)) {
          this.amazonS3Service.deleteImageFromS3Bucket(getBucketName(), user.getAvatarURL());
       }
